Make Id[T] more opaque.
The intention here is to ensure we don't accidentally reuse IDs. By
expanding the counter value to a Long, we eliminate the possibility of
overflow in all reasonable scenarios.

By making Id a non-case class, we remove the ability to .copy() it, as
well as its public accessor and constructor. We also use an atomic
long to ensure each ID has a unique serial value. This means that IDs
can be compared using eq and that there is no way to create a bogus,
duplicate ID.